transfer operation extensions to sonar.cfg
